Itinerarybuilder/requirements.txt
# Core Flask dependencies
Flask==3.0.0
python-dotenv==1.0.0

# Firebase dependencies
firebase-admin==6.2.0

# Authentication & Security
werkzeug==3.0.1

# Geolocation & Maps
geopy==2.4.0
googlemaps==4.10.0

# AI/ML for tagging
sentence-transformers==2.2.2
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1

# Image processing & EXIF
Pillow==10.1.0
ExifRead==3.0.0

# HTTP requests
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# Date/time utilities
python-dateutil==2.8.2

# Logging & utilities
coloredlogs==15.0.1